Fix Linter Errors
