+++ ruby_application/specs/book_spec.rb
@@ -3,7 +3,7 @@ require './classes/book'
 describe Book do
   context 'Test Book class' do
     newbook = Book.new('All prints', '10-10-2021', 'very good')
-    
+
     it 'newbook should be an instance of book' do
       expect(newbook).to be_an_instance_of Book
     end
@@ -19,5 +19,5 @@ describe Book do
     it 'should return false unless @archived or cover_state == bad' do
       expect(newbook.can_be_archived?).to eq false
     end
-end
+  end
 end
